Feature toggle Rekor
This allows for customizing Tekton Chains transparency configuration and
correspondingly the Rekor instance. Three modes are now available:
 * off      - transparency off
 * local    - transparency on, with an in-cluster Rekor instance
 * sigstore - transparency on, with Sigstore public Rekor instance

These are set via ApplicationSet list generator and the custom "rekor"
parameter set there, see argo-cd-apps/base/enterprise-contract.yaml.

It is possible to change this parameter in a gitops manner and the
resulting state should reflect the value of the parameter, e.g. changing
from "local" to "off" should remove the "rekor" ArgoCD Application and
configure the "chains-config" ConfigMap accordingly.

The in-cluster deployed instance of Rekor no longer has an Ingress to
it's service, so it is no longer available outside of the cluster.

The `runAsUser` by default is set to 65533, so the annotations on the
`rekor` namespace are set to allow that. Still when the `rekor-server`
Pod is starting the following is logged:

```
spec:2022/12/20 19:56:39 warning: could not resolve current working directory: stat .: permission denied
```

The creation of`treeID` is forced (`createtree.forced=true`), found no
other way for Rekor not to fail with:

```
api/error.go:72	error processing request	{"requestID": "rekor-server-7fc8f96d54-gkbp6/YAfKPyHOLp-000029", "handler": "tlog.GetLogInfo", "statusCode": 500, "clientMessage": "unexpected error communicating with transparency log", "error": "grpc error: rpc error: code = NotFound desc = tree 1234567890 not found"}
```

This will cause data loss if the `rekor-createtree` Job is re-run,
creating a new `treeID`, and the `rekor-server` Pod is restarted to
reconfigure the new `treeID` from the `rekor-config` ConfigMap.

We can't manage `tekton-chans/chains-config` ConfigMap from the
`pipeline-service` as we need to modify it based on the state of the
Rekor mode from `enterprise-contract` Application, so the
`chains-config` ConfigMap was removed from the `pipeline-service` and
was added to `enterprise-contract` where it can be managed/patched.