Configure a cut-back set of EC rules by default
Collections in EC are a way of specifying a subset of the EC rules
that should be applied. Change the default configuration to specify
that EC should use the "minimal" of rules.

The idea is that a brand new user is not likely to want the full set
of EC checks, which includes some Red Hat specific requirements.

JIRA: HACBS-1615